dotPad: move Dot Pad device enums and constants into defs.py, move driver stuff into driver.py.
