+++ source/brailleDisplayDrivers/dotPad/__init__.py
@@ -1,0 +1,9 @@
+# A part of NonVisual Desktop Access (NVDA)
+# This file is covered by the GNU General Public License.
+# See the file COPYING for more details.
+# Copyright (C) 2024 NV Access Limited
+
+"""Braille display driver for the Dot Pad Tactile Graphic braille display from Dot Inc."""
+
+# Imported here so that braille._getDisplayDriver can import
+from .driver import BrailleDisplayDriver  # noqa: F401
